Melhoria nas explicações em comentário e correção de códigos

+++ 1_C/headers/02_strings_e_caracteres/2_header_string.h/08_sem_biblioteca_string.h/polindromo.c
@@ -12,7 +12,7 @@
 int isPolindrome(char str[])
 {
     int left = 0;
-    int right = strlen(str) - 1;
+    size_t right = strlen(str) - 1;
 
     while (left < right)
     {
